move currently unused uring definitions to a separate file
